requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.11.17
click==8.1.7
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.103.2
greenlet==3.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
iniconfig==2.0.0
mysql-connector-python==8.1.0
mysqlclient==2.2.0
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
protobuf==4.21.12
pyasn1==0.5.0
pydantic==2.4.2
pydantic_core==2.10.1
PyMySQL==1.1.0
pytest==7.4.3
pytest-env==1.1.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
uvicorn==0.23.2
